feat: refactor loop portion of crud to utils
