+++ cloud/src/_utils/index.js
@@ -1,0 +1,7 @@
+const Loop = require('./loop/loop.js');
+
+const utils = {
+    Loop: Loop
+}
+
+module.exports = utils;
